Add method to load data generically on route change.
Instead of each component running its own AJAX on componentDidMount
now we have a fetchData that runs in the `Aff` monad. fetchData takes
a route and returns a route, allowing to e.g. add additional loaded
data or redirection to an error / signin etc.

The remaining changes are warning cleanups for purescript 0.8.