Move remaining error stuff to Internal
I figured out that `APIError` is actually intrinsically part of the
structure of the code -- we use it to be able to write helpers that can
be used in any handler.

Thus, moved APIError to the Internal module and created a new "public"
helper for raising SubAPIErrors, which are the most common.