Fix frontend browsing (for some value of "fix").
Now that we're using repo-ids everywhere except for user-facing code
we can't just proxy requests through to the backend. This is good
because it filters out some crap but it complicates code somewhat.

Mainly we're now parsing & decoding the full response and then sending
that on which introduces slowness (wait for full response, parse,
re-encode, then send first byte).

In the medium term we want API to make the routing decision but then
let repo-server stream its response straight to the http terminator.