Use command-line flags rather than env variables
We want to use flags because:

* they are more explicit, leading to fewer errors
* it is easier to get help to discover how to use a command
* they are already namespaced (`--port` doesn't need a prefix)

It means we can't use `direnv` for funky shell stuff though. Alas.

Updates holborn-repo and holborn-api to use optparse-applicative, which
is a pretty cool library for parsing flags.

In the meantime, we remove the `rawRepoHostname` configuration parameter
from holborn-api, because we only ever use the same server as the core
repo server.

We also begin an as-yet-unfinished transition to calling the ports of
holborn-repo "http port" and "git port" respectively, rather than "port"
and "raw port".