Move form handling from separate Data/Error struct to just one.
We want a "Maybe X" in form data because the elements can be empty
(Nothing) to begin with which is *different* from an empty string ("")
which could be an actual error ("field cannot be empty").

This also removes some reduncancy and room for error by missing fields
in one record that are in the other record.